move value functions to seperate folder
