+++ decision/value_functions/__init__.py
@@ -1,0 +1,1 @@
+from .linear_value_function import LinearValueFunction
